+++ R/select_helpers.R
@@ -2,6 +2,7 @@
 #'
 #' - `all_uppercase()` select all columns with upper case names
 #' - `all_lowercase()` select all columns with lower case names
+#' - `all_labelled()` select all columns with a label attribute
 #' @name select_helpers
 #'
 #' @examples
@@ -9,6 +10,9 @@
 #'
 #' df |> select(all_uppercase())
 #' df |> select(all_lowercase())
+#' df |>
+#'   labelled::set_variable_labels(ONE = "First Column") |>
+#'   select(all_labelled())
 NULL
 
 #' @rdname select_helpers
@@ -23,4 +27,14 @@ all_uppercase <- function() {
 all_lowercase <- function() {
   varnames <- tidyselect::peek_vars()
   which(varnames == tolower(varnames))
+}
+
+#' @rdname select_helpers
+#' @export
+all_labelled <- function() {
+  data <- tidyselect::peek_data()
+
+  which(
+    !lapply(seq_len(ncol(data)), \(x) attr(data[[x]], "label") |> is.null()) |> unlist()
+  )
 }